requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
blas=1.0=mkl
ca-certificates=2020.4.5.2=hecda079_0
certifi=2020.4.5.2=py36h9f0ad1d_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.10.0=py_2
dbus=1.13.6=he372182_0
expat=2.2.9=he1b5a44_2
faiss-cpu=1.6.3=py36h6bb024c_0
fontconfig=2.13.1=he4413a7_1000
freetype=2.9.1=h8a8886c_1
fvcore=0.1.1.post20200610=py_0
glib=2.63.1=h3eb4bd4_1
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb31296c_0
h5py=2.10.0=nompi_py36h513d04c_102
hdf5=1.10.5=nompi_h3c11f04_1104
icu=58.2=hf484d3e_1000
intel-openmp=2020.1=217
joblib=0.15.1=py_0
jpeg=9b=h024ee3a_2
kiwisolver=1.2.0=py36hdb11119_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.3=he6710b0_1
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h14c3975_1000
libxcb=1.13=h14c3975_1002
libxml2=2.9.9=hea5a465_1
lz4-c=1.9.2=he6710b0_0
matplotlib=3.1.3=py36_0
matplotlib-base=3.1.3=py36hef1b27d_0
mkl=2020.1=217
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.1=py36h0573a6f_0
ncurses=6.2=he6710b0_1
ninja=1.9.0=py36hfd86e86_0
numpy=1.18.1=py36h4f9e942_0
numpy-base=1.18.1=py36hde5b4d6_1
olefile=0.46=py36_0
openssl=1.1.1g=h516909a_0
pandas=1.0.3=py36h0573a6f_0
pcre=8.44=he1b5a44_0
pillow=7.1.2=py36hb39fc2d_0
pip=20.0.2=py36_3
portalocker=1.7.0=py36h9f0ad1d_0
pthread-stubs=0.4=h14c3975_1001
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py36hcca6a23_4
python=3.6.10=h7579374_2
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pytorch=1.4.0=py3.6_cuda10.1.243_cudnn7.6.3_0
pytz=2020.1=py_0
pyyaml=5.3.1=py36h8c4c3a4_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
scikit-learn=0.22.1=py36hd81dba3_0
scipy=1.1.0=py36hfa4b5c9_1
setuptools=47.1.1=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.15.0=py_0
sqlite=3.31.1=h62c20be_1
tabulate=0.8.7=pyh9f0ad1d_0
termcolor=1.1.0=py_2
tk=8.6.8=hbc83047_0
torchvision=0.5.0=py36_cu101
tornado=6.0.4=py36h8c4c3a4_1
tqdm=4.46.1=pyh9f0ad1d_0
wheel=0.34.2=py36_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xz=5.2.5=h7b6447c_0
yacs=0.1.6=py_0
yaml=0.2.5=h516909a_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
